Fix type error

+++ src/transformers/generation/logits_process.py
@@ -213,7 +213,7 @@ class EncoderRepetitionPenaltyLogitsProcessor(LogitsProcessor):
     Args:
         hallucination_penalty (`float`):
             The parameter for hallucination penalty. 1.0 means no penalty.
-        encoder_input_ids (`int`):
+        encoder_input_ids (`torch.LongTensor`):
             The encoder_input_ids that should not be repeated within the decoder ids.
     """
